LayoutTests/http/wpt/webauthn/public-key-credential-create-failure-local-silent.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.

PASS PublicKeyCredential's [[create]] with unsupported public key credential parameters in a mock local authenticator. 
PASS PublicKeyCredential's [[create]] with matched exclude credentials in a mock local authenticator. 
PASS PublicKeyCredential's [[create]] with matched exclude credentials in a mock local authenticator. 2nd 
PASS PublicKeyCredential's [[create]] without user consent in a mock local authenticator. 
PASS PublicKeyCredential's [[create]] without private keys in a mock local authenticator. 
PASS PublicKeyCredential's [[create]] without attestation in a mock local authenticator. 
